FIX: subjId and uipatdir as cells in case of BIDS directory ENH: added case in which BIDS root folder with sourcedata of one patient is drag/dropped onto lead-dbs GUI

+++ helpers/ea_legacy2bids.m
@@ -169,8 +169,10 @@ for patients = 1:length(source)
                 if ~isdicom
                     disp("There are no dicom images, source data folder will be empty")
                 else
-                    disp("Copying DICOM folder...");
-                    copyfile(dicom_patient,new_path)                    
+                    if ~exist(dicom_patient, 'dir')
+                        disp("Copying DICOM folder...");
+                        copyfile(dicom_patient,new_path)
+                    end
                 end 
             case 'derivatives'
                 disp("Migrating Derivatives folder...");
